Move checkUIAdminStatus from out of AppContext to AdminContext
